libs/jupyter-deploy-tf-aws-ec2-base/pyproject.toml
[project]
name = "jupyter-deploy-tf-aws-ec2-base"
version = "0.2.0"
description = "Base terraform template to deploy jupyter notebook on AWS EC2."
readme = "README.md"
authors = [
    { name = "Jonathan Guinegagne", email = "jggg@amazon.com" },
    { name = "Michael Chin", email = "chnmch@amazon.com" },
    { name = "Brian Granger", email = "brgrange@amazon.com" }
]
requires-python = ">=3.12"
dependencies = []

[project.entry-points."jupyter_deploy.terraform_templates"]
aws_ec2_base = "jupyter_deploy_tf_aws_ec2_base.template:TEMPLATE_PATH"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["jupyter_deploy_tf_aws_ec2_base"]

[dependency-groups]
dev = [
    "jupyter-deploy",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "python-hcl2>=7.2.1",
    "pyyaml>=6.0.2",
    "ruff>=0.11.8",
    "types-pyyaml>=6.0.12.20250516",
]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
mypy_path = ["."]
packages = ["jupyter_deploy_tf_aws_ec2_base"]

[tool.ruff]
extend-exclude = [".conda/"]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pydocstyle
    "F",  # pyflakes
    "UP",  # pyupgrade
    "B",  # flake8-bugbear
    "SIM",  # flake8-simplify
    "I"   # isort
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

docstring-code-format = true
docstring-code-line-length = 40

[tool.pytest.ini_options]
addopts = "--cov=jupyter_deploy_tf_aws_ec2_base --cov-report=term-missing --cov-report=xml --cov-report=html"
testpaths = ["tests/unit"]

[tool.coverage.run]
source = ["jupyter_deploy_tf_aws_ec2_base"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
fail_under = 80
show_missing = true

[tool.uv.sources]
jupyter-deploy = { workspace = true }
